pyproject.toml
[project]
name = "test-voice-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "livekit-agents[google]~=1.0",
    "livekit-plugins-noise-cancellation~=0.2",
    "pyyaml>=6.0.2",
]
